# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Structured context for the error, or null.",
            "title": "Detail"
          },
          "error_code": {
            "description": "Machine-readable code. Branch on this, not on the message.",
            "enum": [
              "EMPTY_STATEMENT",
              "MULTIPLE_STATEMENTS",
              "NON_SELECT_STATEMENT",
              "FORBIDDEN_KEYWORD",
              "FORBIDDEN_FUNCTION",
              "INVALID_SQL",
              "QUERY_EXECUTION_ERROR",
              "INVALID_CHART_CONFIG",
              "INVALID_CONNECTION_CONFIG",
              "DB_TLS_REQUIRED",
              "CONNECTION_PAUSED",
              "ROW_LIMIT_EXCEEDED",
              "SQL_TOO_LONG",
              "RESULT_TOO_LARGE",
              "WEAK_PASSWORD",
              "DB_AUTH_FAILED",
              "DB_PERMISSION_DENIED",
              "INVALID_CREDENTIALS",
              "NOT_AUTHENTICATED",
              "ACCOUNT_LOCKED",
              "PASSWORD_CHANGE_REQUIRED",
              "FORBIDDEN",
              "CONNECTION_NOT_FOUND",
              "QUERY_NOT_FOUND",
              "TABLE_NOT_FOUND",
              "DASHBOARD_NOT_FOUND",
              "USER_NOT_FOUND",
              "DUPLICATE_NAME",
              "DUPLICATE_EMAIL",
              "LAST_ADMIN",
              "QUERY_FROZEN",
              "REQUEST_VALIDATION_ERROR",
              "RATE_LIMITED",
              "DB_UNREACHABLE",
              "QUERY_TIMEOUT",
              "INTERNAL_ERROR",
              "SERVICE_NOT_READY"
            ],
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error_code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "ChartType": {
        "enum": [
          "line",
          "bar",
          "pie",
          "number",
          "table",
          "compare",
          "movers",
          "compare_grid",
          "heatmap"
        ],
        "title": "ChartType",
        "type": "string"
      },
...
      "DashboardCreate": {
        "properties": {
          "chart_ids": {
            "items": {
              "type": "string"
            },
            "title": "Chart Ids",
            "type": "array"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "DashboardCreate",
        "type": "object"
      },
      "DashboardRead": {
        "properties": {
          "chart_ids": {
            "items": {
              "type": "string"
            },
            "title": "Chart Ids",
            "type": "array"
          },
          "charts": {
            "items": {
              "$ref": "#/components/schemas/QueryChartRead"
            },
            "title": "Charts",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Email"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "owner_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Name"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "chart_ids",
          "created_at",
          "updated_at"
        ],
        "title": "DashboardRead",
        "type": "object"
      },
      "DashboardUpdate": {
        "description": "Every field optional. Omitted fields are left untouched.\n\nPassing ``chart_ids`` replaces the whole arrangement rather than merging\ninto it: a dashboard is an ordered list, and a partial merge has no\nwell-defined meaning for order.",
        "properties": {
          "chart_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chart Ids"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "DashboardUpdate",
        "type": "object"
      },
...
      "QueryChartRead": {
        "properties": {
          "chart_type": {
            "$ref": "#/components/schemas/ChartType"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_public": {
            "title": "Is Public",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "published_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published By"
          },
          "query_id": {
            "title": "Query Id",
            "type": "string"
          },
          "series_field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Field"
          },
          "surge_threshold_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surge Threshold Pct"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "x_field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "X Field"
          },
          "y_field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y Field"
          }
        },
        "required": [
          "id",
          "query_id",
          "name",
          "position",
          "chart_type",
          "x_field",
          "y_field",
          "series_field",
          "surge_threshold_pct",
          "is_public",
          "published_by",
          "published_at",
          "created_at",
          "updated_at"
        ],
        "title": "QueryChartRead",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Schema-agnostic backend for saved, read-only SQL exposed as chart-ready JSON. All fraud logic lives in the SQL you save.",
    "title": "Fraud Analyzer Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/dashboards": {
      "get": {
        "description": "Every dashboard the caller may see, oldest first.",
        "operationId": "list_dashboards_dashboards_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DashboardRead"
                  },
                  "title": "Response List Dashboards Dashboards Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "List Dashboards",
        "tags": [
          "dashboards"
        ]
      },
      "post": {
        "description": "Create a dashboard, optionally populated in one call.\n\nEvery referenced query must exist; nothing is persisted otherwise.",
        "operationId": "create_dashboard_dashboards_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "Create Dashboard",
        "tags": [
          "dashboards"
        ]
      }
    },
    "/dashboards/{dashboard_id}": {
      "delete": {
        "description": "Delete the board. The saved queries it showed are untouched.",
        "operationId": "delete_dashboard_dashboards__dashboard_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "dashboard_id",
            "required": true,
            "schema": {
              "title": "Dashboard Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "Delete Dashboard",
        "tags": [
          "dashboards"
        ]
      },
      "get": {
        "operationId": "get_dashboard_dashboards__dashboard_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dashboard_id",
            "required": true,
            "schema": {
              "title": "Dashboard Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "Get Dashboard",
        "tags": [
          "dashboards"
        ]
      },
      "put": {
        "description": "Rename a dashboard, reorder it, or replace what is on it.\n\n``chart_ids`` replaces the whole arrangement rather than merging into it.",
        "operationId": "update_dashboard_dashboards__dashboard_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "dashboard_id",
            "required": true,
            "schema": {
              "title": "Dashboard Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "Update Dashboard",
        "tags": [
          "dashboards"
        ]
      }
    },
...
  }
}